# Updating some readonly static data in JpegEncoderCore to take advantage of compiler functionality.