@tannergooding 2019-03-21
```
BenchmarkDotNet=v0.11.3, OS=Windows 10.0.17763.379 (1809/October2018Update/Redstone5)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview4-010901
  [Host]     : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
```

Before:

Method |   TestImage |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD |
---------------------- |------------ |---------:|----------:|----------:|---------:|------:|--------:|
'System.Drawing Jpeg' | Bmp/Car.bmp | 4.817 ms | 0.1838 ms | 0.5419 ms | 4.985 ms |  1.00 |    0.00 |
'ImageSharp Jpeg' | Bmp/Car.bmp | 7.079 ms | 0.4893 ms | 1.4428 ms | 6.105 ms |  1.51 |    0.44 |

After:

Method |   TestImage |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD |
---------------------- |------------ |---------:|----------:|----------:|---------:|------:|--------:|
'System.Drawing Jpeg' | Bmp/Car.bmp | 4.857 ms | 0.1933 ms | 0.5698 ms | 5.036 ms |  1.00 |    0.00 |
'ImageSharp Jpeg' | Bmp/Car.bmp | 5.716 ms | 0.0762 ms | 0.0595 ms | 5.716 ms |  1.29 |    0.15 |