@tannergooding 2019-03-21
```
BenchmarkDotNet=v0.11.3, OS=Windows 10.0.17763.379 (1809/October2018Update/Redstone5)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview4-010901
  [Host] : .NET Core 2.1.9 (CoreCLR 4.6.27414.06, CoreFX 4.6.27415.01), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3362.0
```

Before:

Method |   TestImage |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
---------------------- |------------ |---------:|----------:|----------:|---------:|------:|--------:|------------:|------------:|------------:|--------------------:|
'System.Drawing Jpeg' | Bmp/Car.bmp | 4.856 ms | 0.1641 ms | 0.4838 ms | 5.064 ms |  1.00 |    0.00 |    210.9375 |           - |           - |           873.48 KB |
'ImageSharp Jpeg' | Bmp/Car.bmp | 6.041 ms | 0.0559 ms | 0.0467 ms | 6.022 ms |  1.24 |    0.15 |     23.4375 |           - |           - |           129.19 KB |

After:
 
Method |   TestImage |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
---------------------- |------------ |---------:|----------:|----------:|---------:|------:|--------:|------------:|------------:|------------:|--------------------:|
'System.Drawing Jpeg' | Bmp/Car.bmp | 4.909 ms | 0.1418 ms | 0.4180 ms | 5.046 ms |  1.00 |    0.00 |    210.9375 |           - |           - |           873.48 KB |
'ImageSharp Jpeg' | Bmp/Car.bmp | 6.071 ms | 0.1181 ms | 0.1406 ms | 5.995 ms |  1.28 |    0.11 |     23.4375 |           - |           - |           129.19 KB |